# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DARFUS Jewellery ERP API",
    "version": "1.0.0",
    "description": "Production-ready REST API for DARFUS Jewellery ERP"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API Gateway v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/attachments/upload": {
      "post": {
        "summary": "Upload File Attachment",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Attachment uploaded." }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}